refactor: add courses main-logic
